requirements.txt
elasticsearch==7.13.4
elasticsearch_dsl==7.4.0
Flask==2.0.2
numpy==1.21.1
pandas==1.3.4
pymongo==3.12.0
python-dotenv==0.19.2
scikit_learn==1.0.2
gunicorn==20.0.4
psycopg2-binary==2.8.6